It works sans url error testing
So this works to the best of my knowledge. If you look in
BackgroundManagerTest you'll see how to basically do it. At first it
downloads three pictures then it goes into the official run stage. When
you pass more pictures it waits for the first BackgroundManager object
(bm) to finish its run phase and then adds the images onto the existing
array of image urls. It then just picks up where it left off. There are
probably only three lines of code that you need to use to implement
funcitonality (I hope).